add code that could conflict with tanglegram.py and remove vistrans dir to avoid confusion
