--- vistrans/README.md
@@ -1,4 +1,0 @@
-# vistrans
-Visualization code to view reconciliation from eMPRess output, intended for use with Prof Ran and Prof Wu's eMPRess.
-
-Full eMPRess code found here: https://github.com/ssantichaivekin/eMPRess
